Refactor dataframe interchange namespace
